refactor agent code
