+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/advice/CommandExecutorTaskAdvice.java
@@ -23,8 +23,8 @@ import io.opentracing.util.GlobalTracer;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodAroundAdvice;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodInvocationResult;
 import org.apache.shardingsphere.agent.core.plugin.advice.TargetObject;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.ShardingErrorSpan;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingTags;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.span.ErrorSpan;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingSphereTags;
 import org.apache.shardingsphere.infra.executor.kernel.model.ExecutorDataMap;
 
 import java.lang.reflect.Method;
@@ -32,7 +32,7 @@ import java.lang.reflect.Method;
 /**
  * Command executor task advice.
  */
-public class CommandExecutorTaskAdvice implements MethodAroundAdvice {
+public final class CommandExecutorTaskAdvice implements MethodAroundAdvice {
     
     private static final String OPERATION_NAME = "/ShardingSphere/rootInvoke/";
     
@@ -41,7 +41,7 @@ public class CommandExecutorTaskAdvice implements MethodAroundAdvice {
     @Override
     public void beforeMethod(final TargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
         Scope scope = GlobalTracer.get().buildSpan(OPERATION_NAME)
-                .withTag(Tags.COMPONENT.getKey(), ShardingTags.COMPONENT_NAME)
+                .withTag(Tags.COMPONENT.getKey(), ShardingSphereTags.COMPONENT_NAME)
                 .startActive(true);
         ExecutorDataMap.getValue().put(ROOT_SPAN, scope.span());
     }
@@ -54,6 +54,6 @@ public class CommandExecutorTaskAdvice implements MethodAroundAdvice {
     
     @Override
     public void onThrowing(final TargetObject target, final Method method, final Object[] args, final Throwable throwable) {
-        ShardingErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
+        ErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
     }
 }
