+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/advice/JDBCExecutorCallbackAdvice.java
@@ -25,8 +25,8 @@ import io.opentracing.util.GlobalTracer;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodAroundAdvice;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodInvocationResult;
 import org.apache.shardingsphere.agent.core.plugin.advice.TargetObject;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.ShardingErrorSpan;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingTags;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.span.ErrorSpan;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingSphereTags;
 import org.apache.shardingsphere.infra.executor.sql.context.ExecutionUnit;
 import org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutionUnit;
 
@@ -36,7 +36,7 @@ import java.util.Map;
 /**
  * JDBC executor callback advice.
  */
-public class JDBCExecutorCallbackAdvice implements MethodAroundAdvice {
+public final class JDBCExecutorCallbackAdvice implements MethodAroundAdvice {
     
     private static final String OPERATION_NAME = "/ShardingSphere/executeSQL/";
     
@@ -49,12 +49,12 @@ public class JDBCExecutorCallbackAdvice implements MethodAroundAdvice {
         } else {
             JDBCExecutionUnit executionUnit = (JDBCExecutionUnit) args[0];
             ExecutionUnit unit = executionUnit.getExecutionUnit();
-            builder.withTag(Tags.COMPONENT.getKey(), ShardingTags.COMPONENT_NAME)
+            builder.withTag(Tags.COMPONENT.getKey(), ShardingSphereTags.COMPONENT_NAME)
                     .withTag(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_CLIENT)
                     .withTag(Tags.DB_TYPE.getKey(), "sql")
                     .withTag(Tags.DB_INSTANCE.getKey(), unit.getDataSourceName())
                     .withTag(Tags.DB_STATEMENT.getKey(), unit.getSqlUnit().getSql())
-                    .withTag(ShardingTags.DB_BIND_VARIABLES.getKey(), unit.getSqlUnit().getParameters().toString());
+                    .withTag(ShardingSphereTags.DB_BIND_VARIABLES.getKey(), unit.getSqlUnit().getParameters().toString());
         }
         target.setAttachment(builder.startActive(true));
     }
@@ -66,6 +66,6 @@ public class JDBCExecutorCallbackAdvice implements MethodAroundAdvice {
     
     @Override
     public void onThrowing(final TargetObject target, final Method method, final Object[] args, final Throwable throwable) {
-        ShardingErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
+        ErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
     }
 }
