+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/advice/SQLParserEngineAdvice.java
@@ -23,22 +23,22 @@ import io.opentracing.util.GlobalTracer;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodAroundAdvice;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodInvocationResult;
 import org.apache.shardingsphere.agent.core.plugin.advice.TargetObject;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.ShardingErrorSpan;
-import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingTags;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.span.ErrorSpan;
+import org.apache.shardingsphere.agent.plugin.tracing.jaeger.constant.ShardingSphereTags;
 
 import java.lang.reflect.Method;
 
 /**
  * SQL parser engine advice.
  */
-public class SQLParserEngineAdvice implements MethodAroundAdvice {
+public final class SQLParserEngineAdvice implements MethodAroundAdvice {
     
     private static final String OPERATION_NAME = "/ShardingSphere/parseSQL/";
     
     @Override
     public void beforeMethod(final TargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
         Scope scope = GlobalTracer.get().buildSpan(OPERATION_NAME)
-                .withTag(Tags.COMPONENT.getKey(), ShardingTags.COMPONENT_NAME)
+                .withTag(Tags.COMPONENT.getKey(), ShardingSphereTags.COMPONENT_NAME)
                 .withTag(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_CLIENT)
                 .withTag(Tags.DB_STATEMENT.getKey(), String.valueOf(args[0]))
                 .startActive(true);
@@ -52,6 +52,6 @@ public class SQLParserEngineAdvice implements MethodAroundAdvice {
     
     @Override
     public void onThrowing(final TargetObject target, final Method method, final Object[] args, final Throwable throwable) {
-        ShardingErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
+        ErrorSpan.setError(GlobalTracer.get().activeSpan(), throwable);
     }
 }
