+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-jaeger/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/jaeger/service/JaegerTracerBootService.java
@@ -15,43 +15,42 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.agent.plugin.tracing.zipkin;
+package org.apache.shardingsphere.agent.plugin.tracing.jaeger.service;
 
-import brave.Tracing;
+import io.jaegertracing.Configuration;
+import io.opentracing.util.GlobalTracer;
 import org.apache.shardingsphere.agent.core.config.AgentConfiguration;
 import org.apache.shardingsphere.agent.core.plugin.service.BootService;
 import org.apache.shardingsphere.agent.core.cache.AgentObjectPool;
-import zipkin2.reporter.brave.AsyncZipkinSpanHandler;
-import zipkin2.reporter.okhttp3.OkHttpSender;
 
 /**
- * Zipkin tracer boot service.
+ * Jaeger tracer boot service.
  */
-public final class ZipkinTracerBootService implements BootService {
+public final class JaegerTracerBootService implements BootService {
     
-    private AsyncZipkinSpanHandler zipkinSpanHandler;
-    
-    private OkHttpSender sender;
-    
-    private Tracing tracing;
-
     @Override
     public void setup() {
         AgentConfiguration configuration = AgentObjectPool.INSTANCE.get(AgentConfiguration.class);
         AgentConfiguration.TracingConfiguration tracingConfiguration = configuration.getTracing();
-        sender = OkHttpSender.create("http://" + tracingConfiguration.getAgentHost() + ":" + tracingConfiguration.getAgentPort());
-        zipkinSpanHandler = AsyncZipkinSpanHandler.create(sender);
+        tracingConfiguration.getExtra().forEach(System::setProperty);
+        Configuration.SamplerConfiguration samplerConfig = Configuration.SamplerConfiguration.fromEnv();
+        Configuration.ReporterConfiguration reporterConfig = Configuration.ReporterConfiguration.fromEnv()
+                .withSender(
+                        Configuration.SenderConfiguration.fromEnv()
+                                .withAgentHost(tracingConfiguration.getAgentHost())
+                                .withAgentPort(tracingConfiguration.getAgentPort())
+                );
+        Configuration config = new Configuration(configuration.getApplicationName())
+                .withSampler(samplerConfig)
+                .withReporter(reporterConfig);
+        GlobalTracer.register(config.getTracer());
     }
-
+    
     @Override
     public void start() {
-        tracing = Tracing.newBuilder().localServiceName("shardingsphere-agent").addSpanHandler(zipkinSpanHandler).build();
     }
-
+    
     @Override
     public void cleanup() {
-        tracing.close();
-        zipkinSpanHandler.close();
-        sender.close();
     }
 }
