+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-zipkin/pom.xml
@@ -32,7 +32,7 @@
     <properties>
         <zipkin-reporter.version>2.16.2</zipkin-reporter.version>
         <zipkin-brave.version>5.13.2</zipkin-brave.version>
-        <entrypoint.class>org.apache.shardingsphere.agent.plugin.tracing.zipkin.ZipkinPluginDefinition</entrypoint.class>
+        <entrypoint.class>org.apache.shardingsphere.agent.plugin.tracing.zipkin.definition.ZipkinPluginDefinition</entrypoint.class>
     </properties>
 
     <dependencies>
