+++ shardingsphere-agent/shardingsphere-agent-plugins/shardingsphere-agent-plugin-tracing/shardingsphere-agent-tracing-zipkin/src/main/java/org/apache/shardingsphere/agent/plugin/tracing/zipkin/advice/CommandExecutorTaskAdvice.java
@@ -23,7 +23,7 @@ import lombok.SneakyThrows;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodAroundAdvice;
 import org.apache.shardingsphere.agent.core.plugin.advice.MethodInvocationResult;
 import org.apache.shardingsphere.agent.core.plugin.advice.TargetObject;
-import org.apache.shardingsphere.agent.plugin.tracing.zipkin.ShardingConstants;
+import org.apache.shardingsphere.agent.plugin.tracing.zipkin.constant.ZipkinConstants;
 import org.apache.shardingsphere.infra.executor.kernel.model.ExecutorDataMap;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.frontend.command.CommandExecutorTask;
@@ -33,29 +33,29 @@ import java.lang.reflect.Method;
 /**
  * Command executor task advice.
  */
-public class CommandExecutorTaskAdvice implements MethodAroundAdvice {
+public final class CommandExecutorTaskAdvice implements MethodAroundAdvice {
 
     private static final String OPERATION_NAME = "/ShardingSphere/rootInvoke/";
 
     @Override
     public void beforeMethod(final TargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
         Span span = Tracing.currentTracer().newTrace().name(OPERATION_NAME);
-        span.tag(ShardingConstants.Tags.COMPONENT, ShardingConstants.COMPONENT_NAME).kind(Span.Kind.CLIENT)
-                .tag(ShardingConstants.Tags.DB_STATEMENT, ShardingConstants.DB_TYPE_VALUE).start();
-        ExecutorDataMap.getValue().put(ShardingConstants.ROOT_SPAN, span);
+        span.tag(ZipkinConstants.Tags.COMPONENT, ZipkinConstants.COMPONENT_NAME).kind(Span.Kind.CLIENT)
+                .tag(ZipkinConstants.Tags.DB_STATEMENT, ZipkinConstants.DB_TYPE_VALUE).start();
+        ExecutorDataMap.getValue().put(ZipkinConstants.ROOT_SPAN, span);
     }
 
     @SneakyThrows
     @Override
     public void afterMethod(final TargetObject target, final Method method, final Object[] args, final MethodInvocationResult result) {
         BackendConnection connection = (BackendConnection) CommandExecutorTask.class.getDeclaredField("backendConnection").get(target);
-        Span span = (Span) ExecutorDataMap.getValue().remove(ShardingConstants.ROOT_SPAN);
-        span.tag(ShardingConstants.Tags.CONNECTION_COUNT, String.valueOf(connection.getConnectionSize()));
+        Span span = (Span) ExecutorDataMap.getValue().remove(ZipkinConstants.ROOT_SPAN);
+        span.tag(ZipkinConstants.Tags.CONNECTION_COUNT, String.valueOf(connection.getConnectionSize()));
         span.flush();
     }
 
     @Override
     public void onThrowing(final TargetObject target, final Method method, final Object[] args, final Throwable throwable) {
-        ((Span) ExecutorDataMap.getValue().get(ShardingConstants.ROOT_SPAN)).error(throwable);
+        ((Span) ExecutorDataMap.getValue().get(ZipkinConstants.ROOT_SPAN)).error(throwable);
     }
 }
